chore: move components of performance analytics to corresponding folder
