+++ apps/frontend-manage/src/components/analytics/performance/ElementFeedbackBarChart.tsx
